requirements.txt
fastapi==0.115.0
Flask==3.0.3
protobuf>=3.19.5,<5
python-dotenv==1.0.1
uvicorn==0.31.1
google-generativeai==0.5.1